refactor(Worker): Extract common requests settings
